+++ components/WishlistItem.tsx
@@ -1,4 +1,8 @@
-import { WishlistItemType } from '@/lib/types'
+import { WishlistItemType, User } from '@/lib/types'
+import { useAtom } from 'jotai'
+import { usersAtom } from '@/lib/atoms'
+import { Avatar, AvatarFallback, AvatarImage } from './ui/avatar'
+import { useHelpers } from '@/lib/client-helpers'
 import { Card, CardContent, CardDescription, CardFooter, CardHeader, CardTitle } from '@/components/ui/card'
 import ReactMarkdown from 'react-markdown'
 import { Button } from '@/components/ui/button'
@@ -24,6 +28,25 @@ interface WishlistItemProps {
   isArchived?: boolean
 }
 
+const renderUserAvatars = (item: WishlistItemType, currentUser: User | null, usersData: { users: User[] }) => {
+  if (!item.userIds || item.userIds.length <= 1) return null;
+  
+  return (
+    <div className="flex -space-x-2 ml-2 flex-shrink-0">
+      {item.userIds?.filter((u) => u !== currentUser?.id).map(userId => {
+        const user = usersData.users.find(u => u.id === userId)
+        if (!user) return null
+        return (
+          <Avatar key={user.id} className="h-6 w-6">
+            <AvatarImage src={user?.avatarPath && `/api/avatars/${user.avatarPath.split('/').pop()}` || ""} />
+            <AvatarFallback>{user.username[0]}</AvatarFallback>
+          </Avatar>
+        )
+      })}
+    </div>
+  );
+};
+
 export default function WishlistItem({
   item,
   onEdit,
@@ -35,6 +58,9 @@ export default function WishlistItem({
   isHighlighted,
   isRecentlyRedeemed
 }: WishlistItemProps) {
+  const { currentUser } = useHelpers()
+  const [usersData] = useAtom(usersAtom)
+  
   return (
     <Card
       id={`wishlist-${item.id}`}
@@ -53,11 +79,16 @@ export default function WishlistItem({
             </span>
           )}
         </div>
-        {item.description && (
-          <CardDescription className={`whitespace-pre-line ${item.archived ? 'text-gray-400 dark:text-gray-500' : ''}`}>
-            {item.description}
-          </CardDescription>
-        )}
+        <div className="flex items-center justify-between">
+          <div className="flex-1">
+            {item.description && (
+              <CardDescription className={`whitespace-pre-line ${item.archived ? 'text-gray-400 dark:text-gray-500' : ''}`}>
+                {item.description}
+              </CardDescription>
+            )}
+          </div>
+          {renderUserAvatars(item, currentUser as User, usersData)}
+        </div>
       </CardHeader>
       <CardContent className="flex-1">
         <div className="flex items-center gap-2">